+++ auto_route/lib/src/auto_route_wrapper.dart
@@ -1,4 +1,4 @@
-import 'package:flutter/material.dart';
+import 'package:flutter/material.dart' show BuildContext, Widget;
 
 // clients will implement this class to provide a wrapped route.
 abstract class AutoRouteWrapper {