+++ example/lib/main.dart
@@ -1,8 +1,8 @@
 import 'package:auto_route/auto_route.dart';
+import 'package:example/router/route_guards.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
-import 'router/route_guards.dart';
 import 'router/router.gr.dart';
 
 void main() {
@@ -17,14 +17,13 @@ class MyApp extends StatelessWidget {
       // the state of ExtendedNavigator so it won't reload
       // when using Flutter tools-> select widget mode
       builder: ExtendedNavigator.builder(
-        router: Router(),
-        initialRoute: "/",
-        guards: [AuthGuard()],
-        builder: (_, extendedNav) => Theme(
-          data: ThemeData(brightness: Brightness.dark),
-          child: extendedNav,
-        ),
-      ),
+          router: Router(),
+          initialRoute: '/',
+          guards: [AuthGuard()],
+          builder: (ctx, extendedNav) => Theme(
+                data: ThemeData.dark(),
+                child: extendedNav,
+              )),
     );
   }
 }