+++ example/lib/router/route_guards.dart
@@ -1,13 +1,17 @@
 import 'package:auto_route/auto_route.dart';
+
 import 'router.gr.dart';
+
 // guard
 class AuthGuard extends RouteGuard {
+  @override
   Future<bool> canNavigate(
     ExtendedNavigatorState navigator,
     String routeName,
     Object arguments,
   ) async {
-    return navigator.root.replace(Routes.loginScreen);
+    // if isAuthenticated return true;
+    // else
+    return navigator.root.push(Routes.loginScreen);
   }
 }
-