+++ auto_route/CHANGELOG.md
@@ -1,5 +1,10 @@
 # ChangeLog
-
+## [0.6.7] 
+- Fix blue screen issue when using guards on initial route #189
+- Fix relative import issues
+- Add option to not prefer relative imports 
+- Fix some linter warnings
+- Fix issue #176
 ## [0.6.6] 
 - Fix hot reload issue #170
 - Fix Hero animations not working with the new ExtendedNavigator # 169